[hdf5] move our hdf5 c++ wrapper to h5x/ subdir
